+++ wsei-todo-app-master/src/actions/todolistActions.ts
@@ -1,0 +1,7 @@
+import * as actionTypes from '../actions/types/todolistTypes';
+import { ISingleElList } from '../entitles/todoSingleEl';
+
+export const setNewElTodoList = (newElement: ISingleElList) => ({
+    type: actionTypes.SET_NEW_ELEMENT,
+    newElement
+})