+++ wsei-todo-app-master/CHANGELOG.md
@@ -38,3 +38,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
     - -----------------
 
+## [1.1.0] - 2020-04-20
+
+### Changes
+
+    - TodoList structure 
+    - Remove Notifications, Contacts
+
+### Added
+
+    - Account page file